+++ packages/editor/CHANGELOG.md
@@ -12,6 +12,11 @@
 - Block Settings Section: improve design and UX.
 - Add `&` support for targeting current block in custom CSS feature.
 - `Flex Child` settings support added to `Column` block.
+- Compatibility for detecting variables from core blocks if the variable not found.
+
+### Automated Tests
+- Automated test to check color WP data compatibility if variables not found. 
+
 
 ## 2.0.1 (2025-07-22)